python/pyproject.toml
[project]
name = "scitq2"
version = "c49f199-dirty"
description = "Python DSL for defining scitq2 workflows"
authors = [{ name="Raynald de Lahondès" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0"
]

[project.urls]
"Homepage" = "https://github.com/scitq/scitq"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

